load_function/requirements.txt
functions-framework==3.3.0
google-cloud-bigquery==3.4.0
numpy==1.23.5
